uint8

+++ src/utils/read_conversion.py
@@ -51,7 +51,7 @@ def file_to_matrix(filename="/tmp/outfile.txt", r=180, n_reads=None):
         sys.stdout.write("Number of reads not specified , estimating the number\n")
         estimated_reads = input_file_size/avg_record_bytes
     sys.stdout.write("Estimated/specified number of reads: {}\n".format(estimated_reads))
-    reads_matrix = ndarray(shape=(estimated_reads, r), dtype=np.uint32)
+    reads_matrix = ndarray(shape=(estimated_reads, r), dtype=np.uint8)
     reads_matrix[0] = map(base_to_int,list(str(second_line)))
     counter = 1
     in_file.readline()  # Skip id line
